requirements.txt
stable-baselines3
swig
gym
gym[box2d]
tensorboard // optional. You can also just use wandb
#opencv-python // Only used for the tutorial part?
#JSSEnv // Copied into my own envs folder
seaborn
ipython
sb3-contrib
wandb
pytest
pytest-sugar
pytest-cov
sympy
sklearn
torchvision
tqdm
plotly
tabulate
ipykernel
jupyterlab
notebook
ortools
chart_studio
python-dotenv